# TPE Arquitectura Web Final/openapi.yaml
openapi: 3.0.1
info:
  title: API Monopatines Electricos
  description: Esta API se encarga del manego de un sistema de monopatines electricos
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /at/tarifas/:
    get:
      tags:
      - Tarifas
      summary: obtiene las tarifas
      responses: 
        200: 
          description: (ok) la peticion fue exitosa
          content: 
            aplication/json: 
              schema:
                $ref: '#/components/schemas/Tarifa'
        400: 
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /at/tarifas/id/{id}:
    get:
      tags:
      - Tarifas
      summary: obtiene la tarifas dado un id
      parameters:
      - name: id
        in: path  
        required: true
        schema:
          type: string
      responses: 
        200: 
          description: (ok) la peticion fue exitosa
          content: 
            aplication/json: 
              schema:
                $ref: '#/components/schemas/Tarifa'
        400: 
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /at/tarifas/agregar:
     post:
      tags:
      - Tarifas
      summary: agrega otra tarifa
      requestBody: 
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/Tarifa'
              
      responses: 
        200: 
          description: (ok) la peticion fue exitosa
          content: 
            aplication/json: 
              schema:
                $ref: '#/components/schemas/Administrador'
        400: 
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
  /p/paradas:
    get:
        tags:
        - Paradas
        summary: obtiene las paradas
        responses: 
          200: 
            description: (ok) la peticion fue exitosa
          400: 
            $ref: '#/components/responses/BadRequest'
          401:
            $ref: '#/components/responses/Unauthorized'
          404:
            $ref: '#/components/responses/NotFound'
          500:
            $ref: '#/components/responses/ServerError'
  /uc/cuentas/:
    get:
        tags:
        - Cuentas
        summary: obtiene las cuentas
        responses: 
          200: 
            description: (ok) la peticion fue exitosa
          400: 
            $ref: '#/components/responses/BadRequest'
          401:
            $ref: '#/components/responses/Unauthorized'
          404:
            $ref: '#/components/responses/NotFound'
          500:
            $ref: '#/components/responses/ServerError'
  /uc/usuarios/:
    get:
        tags:
        - Usuarios
        summary: obtiene los usuarios
        responses: 
          200: 
            description: (ok) la peticion fue exitosa
          400: 
            $ref: '#/components/responses/BadRequest'
          401:
            $ref: '#/components/responses/Unauthorized'
          404:
            $ref: '#/components/responses/NotFound'
          500:
            $ref: '#/components/responses/ServerError'
...
components: 
  responses: 
    Unauthorized:
      description: (Unauthorized) no tienes autorizacion 
    NotFound:
      description: (NotFound) elemento no encontrado
    BadRequest:
      description: (BadRequest) datos invalidos
    ServerError:
      description: (ServerError) error en el servidor
  schemas:
    Administrador:
      type: object
      properties:
        nombre_usuario:
          type: string
        contraseña:
          type: string
    Tarifa:
      type: object
      properties:
        fecha_inicio:
          type: string
        tarifa_normal:
          type: number
        tarifa_extra:
          type: number